removing changes on the input value. only keeping onscreen keyboard changes
